finish libp2p init, fix tests
