+++ harmony/harmonydb/sql/20240228-piece-park.sql
@@ -34,7 +34,7 @@ create table parked_piece_refs (
     data_headers jsonb not null default '{}',
 
     -- host Added in 202240730-market-migrations.sql
-    host text,
+    -- host text,
 
     foreign key (piece_id) references parked_pieces(id) on delete cascade
 );
