+++ tasks/gc/pipeline_meta_gc.go
@@ -158,7 +158,7 @@ func (s *PipelineGC) cleanupMK12DealPipeline() error {
 	ctx := context.Background()
 
 	// Execute the query
-	_, err := s.db.Exec(ctx, `DELETE FROM market_mk12_deal_pipeline WHERE sealed = TRUE AND indexed = TRUE;`)
+	_, err := s.db.Exec(ctx, `DELETE FROM market_mk12_deal_pipeline WHERE complete = TRUE;`)
 	if err != nil {
 		return xerrors.Errorf("failed to clean up sealed deals: %w", err)
 	}
